refactoe(details panel) refactor mobile view, new button and create new widgets
